# server/openapi.yaml
openapi: 3.0.3
info:
  title: Elytra API
  version: 1.0.0
  description: |
    REST API for the Elytra platform.

    - `/public/auth/*`: public authentication flows (sign-up, login, password reset, token refresh).
    - `/private/*`: requires `Authorization: Bearer <idToken>` from login or refresh.
    - `/dev/*`: developer tools, available on the dev and qa stages (and locally), never on prod.

    All paths are relative to the server base URL (e.g. `/api/public/auth/login`).

    **Envelope.** A success response is `{ "data": … }`; an error response is `{ "message": "…" }`.
    There is no `success` field; use the HTTP status code.

    **Files never travel through the API.** Anything sent to `/api` is capped by the Lambda
    payload limit (6,291,556 bytes, about 4.7 MB of file after base64) and API Gateway's 30 s
    timeout. Uploads go straight to S3 with a presigned POST (`/private/uploads/presign`), and
    downloads come from S3 through a presigned GET (`/private/me/export`).
servers:
  - url: http://localhost:3000/api
    description: Local development
  - url: https://{domain}/api
    description: Deployed environment
    variables:
      domain:
        default: elytra.shalev396.com
...
security: []
paths:
...
  /private/me:
    get:
      tags: [User]
      summary: Get current user profile
      description: Returns the authenticated user's profile.
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [User]
      summary: Update current user profile
      description: |
        Updates the name and/or photo. A new photo is first uploaded to S3 through
        `POST /private/uploads/presign` (purpose `account-photo`) and then referenced here by its
        `stagingKey`. The server checks the staged object's owner, size, type and magic bytes,
        strips its metadata, and only then replaces the current photo; a rejected upload leaves
        the current photo untouched. The staged object is consumed, so replaying the same
        `stagingKey` returns 404.

        `removePhoto: true` deletes the current photo without a replacement. A body that changes
        nothing returns 400 "No changes provided".
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequestBody'
            examples:
              name:
                summary: Change the name
                value:
                  name: Jane Doe
              photo:
                summary: Replace the photo with a staged upload
                value:
                  photo:
                    stagingKey: tmp/staging/2b5d0a52-4bd0-4f1a-9a2e-6c1f0e7b6f10/7f1c1d64-1f0e-4a3a-9c6f-2f4b8d0a1b23.png
                    fileName: me.png
              removePhoto:
                summary: Remove the photo
                value:
                  removePhoto: true
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Invalid body, no changes, invalid staging key, disallowed type, or content that doesn't match its type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                message: No changes provided
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found, or the staged upload does not exist (expired, never uploaded, or already consumed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                message: Staged upload not found. Request a new upload URL.
        '413':
          description: The staged file is larger than the purpose allows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              example:
                message: File too large. Maximum is 5 MB.
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /private/me/export:
    get:
      tags: [User]
      summary: Export account data
      description: |
        Builds a ZIP of the user's data (`user-data.csv` plus uploaded files under `assets/`),
        stores it in S3 and returns a presigned download URL, valid for 15 minutes. The object
        is deleted automatically after a day. The download sets
        `Content-Disposition: attachment; filename="<filename>"`.
      operationId: exportMyData
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Presigned download URL for the export ZIP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportMyDataResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /private/delete:
    delete:
      tags: [User]
      summary: Delete user account
      description: Permanently deletes the user's uploads, database records and Cognito user.
      operationId: deleteAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito-issued JWT id token (from login or refresh)

  # ── Reusable Responses ──────────────────────────────────────────────────────
  responses:
...
    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: No authorization header provided
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: User not found
    TooManyRequests:
      description: Rate limit exceeded (per IP, 15-minute window; `RateLimit-*` headers describe the budget)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: Too many requests. Please try again later.
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            message: An unexpected error occurred

  # ── Schemas ─────────────────────────────────────────────────────────────────
  schemas:
...
    ApiErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    MeResponseData:
      type: object
      required: [id, cognitoSub, email, name, photoUrl, lastLoginAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          description: Database id
        cognitoSub:
          type: string
        email:
          type: string
        name:
          type: string
        photoUrl:
          type: string
          format: uri
          nullable: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MeResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/MeResponseData'
    StagedFileReference:
      type: object
      required: [stagingKey, fileName]
      properties:
        stagingKey:
          type: string
          description: From POST /private/uploads/presign
        fileName:
          type: string
          description: Original file name, kept on the media record and used in the export
    UpdateMeRequestBody:
      type: object
      properties:
        name:
          type: string
        removePhoto:
          type: boolean
          description: Delete the current photo without a replacement. Ignored when `photo` is present.
        photo:
          $ref: '#/components/schemas/StagedFileReference'
...
    ExportMyDataResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [downloadUrl, filename]
          properties:
            downloadUrl:
              type: string
              format: uri
              description: Presigned S3 GET, valid for 15 minutes
            filename:
              type: string
              example: user-export-1789867358634.zip
...
    DeleteUserResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [message]
          properties:
            message:
              type: string

    # ── Uploads ───────────────────────────────────────────────────────────────
...